feat(Base): Merged Flex into the Base component.
All of the props that were available on Flex are now available on Base, the only prop name to change is "direction" which has now become "flex".

BREAKING CHANGE: Removed the Flex component. Any uses should be replaced with the Base component, and the "direction" prop should be replaced with "flex".